pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "sec-insights"
version = "0.1.0"
description = "A RAG-based system for querying SEC 10-K filings"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "sec_insights", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
python-dotenv = "^1.0.1"
qdrant-client = {extras = ["fastembed"], version = "^1.9.2"}
openai = "^1.35.3"
pandas = "<2.0"
scikit-learn = "^1.5.0"
numpy = "^1.26.4"
matplotlib = "^3.9.0"
seaborn = "^0.13.2"
jupyter = "^1.0.0"
ipywidgets = "^8.1.3"
tenacity = "^8.4.1"
backoff = "^2.2.1"
rouge-score = "^0.1.2"
tqdm = "^4.66.4"
tiktoken = "^0.7.0"
torch = "^2.3.1"
transformers = "^4.41.2"
sentence-transformers = "^4.1.0"
einops = "^0.8.1"
umap = "^0.1.1"
plotly = "^6.1.2"
umap-learn = "^0.5.7"
statsmodels = "^0.14.4"
pygam = "^0.9.1"
evaluate = "^0.4.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
ruff = "^0.4.8"

[tool.ruff]
line-length = 120
exclude = ["notebooks/"]

[tool.isort]
profile = "black"
skip = ["notebooks"]

[tool.mypy]
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = "src.evaluation.evaluator"
disable_error_code = ["assignment"]
